+++ src/maths/ProbabilityDistribution.java
@@ -2,5 +2,5 @@ package maths;
 
 
 public interface ProbabilityDistribution {	
-	public int getSample();
+	public double getSample();
 }
